UBUNTU: [Config] do not enforce CONFIG_ARM64_BTI_KERNEL
CONFIG_ARM64_BTI_KERNEL depends on a newer gcc version than it's
available in Focal, so this config gets removed when 'make' is run in a
Focal chroot and 'GCC_VERSION' is updated.

The config check started to fail in linux-hwe-5.8 after the following
commit:
"UBUNTU: [Config] Update annotations to include missing options".

Remove the enforcement from the annotations file.

Ignore: yes
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>