UBUNTU: [Debian] run install-$(flavour) targets during build phase
BugLink: https://bugs.launchpad.net/bugs/1930713
BugLink: https://bugs.launchpad.net/bugs/1939287

Move installation of files from the flavour build directories to the
build phase. This results in cleaning up of one flavour build
directory before starting the build of the next flavour, significantly
reducing the amount of space needed on builders.

Note that this will result in incorrect ownership of files in cases
where the build and binary phases of building packages are run
separately. This will be addressed in a later commit.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>