UBUNTU: [Packaging] Move VM DRM drivers into modules
BugLink: https://bugs.launchpad.net/bugs/1960633

We already have vboxvideo included in the linux-modules package for
newer releases. This is also included when using the linux-virtual
meta. But there is a couple of other DRM drivers which could be used
by various VM installations which run a desktop:

- bochs-drm (? Qemu)
- cirrus (old standard when it comes to emulated graphics)
- drm_xen_front (Xen)
- virtio-gpu (KVM)
- vboxvideo (VirtualBox)
- vmwgfx (VMWare)

All of those drivers direct dependencies are already part of
linux-modules. For Focal this also adds vboxvideo which until
now was part of linux-modules-extra.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>